Locationを10個に制限 Locationの緯度・経度・高度の値をdoubleに変更 Locationをjsonで出力

+++ fuel/app/migrations/003_create_locations.php
@@ -9,9 +9,9 @@ class Create_locations
 		\DBUtil::create_table('locations', array(
 			'id' => array('constraint' => 11, 'type' => 'int', 'auto_increment' => true),
 			'pass_id' => array('constraint' => 11, 'type' => 'int'),
-			'latitude' => array('type' => 'float'),
-			'longitude' => array('type' => 'float'),
-			'altitude' => array('type' => 'float', 'null' => true),
+			'latitude' => array('type' => 'double'),
+			'longitude' => array('type' => 'double'),
+			'altitude' => array('type' => 'double', 'null' => true),
 			'relevant_text' => array('constraint' => 255, 'type' => 'varchar', 'null' => true),
 			'created_at' => array('constraint' => 11, 'type' => 'int'),
 			'updated_at' => array('constraint' => 11, 'type' => 'int'),
